Q1 Statement
An egg distribution company uses different sizes of packings for eggs, that is, 30 eggs
packing, 24 eggs packing, 18 eggs packing, 12 eggs packing and 6 eggs packing. Write a program which
prompts user to enter total number of eggs to be packed and then calculate how many packings of
each size will be possible. Also tell if there will be any eggs left to be packed.